requirements.txt
# local package
# -----------------------
# Core Python utilities
# -----------------------
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
joblib>=1.3

# -----------------------
# Machine Learning
# -----------------------
scikit-learn>=1.3

# -----------------------
# DVC & Experiment Tracking
# -----------------------
dvc>=3.0
dvclive>=3.0

# -----------------------
# Visualization (optional but useful)
# -----------------------
matplotlib>=3.7
seaborn>=0.12

# -----------------------
# Jupyter (for notebooks)
# -----------------------
jupyter
ipykernel

# -----------------------
# MLflow (optional – keep if you use it)
# -----------------------
mlflow>=2.9

